feat(sdk): add @dawn-ai/sdk/testing as canonical home for test helpers
Move expectError, expectMeta, expectOutput and the RuntimeExecutionResult
type family to @dawn-ai/sdk/testing. The cli now re-exports from sdk;
@dawn-ai/cli/testing keeps working as a deprecated back-compat path.

This closes the long-standing F-082 confusion (audit #39) where every
doc reader assumed the helpers were at sdk/testing. They are now.

- packages/sdk/src/testing/index.ts — new home for the helpers
- packages/sdk/src/runtime-result.ts — moved type family from cli
- packages/sdk/package.json — adds ./testing subpath export
- packages/cli/src/lib/runtime/result.ts — re-exports types from sdk,
  keeps the cli-local constructor functions
- packages/cli/src/testing/index.ts — re-exports from sdk with @deprecated
- packages/cli/src/commands/test.ts — imports from sdk/testing
- packages/cli/src/lib/runtime/assertions.ts — deleted (no internal callers)
- vitest configs (cli, langchain, test/runtime) — add sdk/testing alias
- docs (testing.mdx, AGENTS.md, CLAUDE.md) — point to sdk/testing
- harness + handwritten fixture — use sdk/testing
- cli-testing-export.test.ts — exercises BOTH subpaths via describe.each

312 tests passing. Lint clean. Docs check clean.
