+++ packages/cli/src/testing/index.ts
@@ -1,1 +1,6 @@
-export { expectError, expectMeta, expectOutput } from "../lib/runtime/assertions.js"
+/**
+ * @deprecated Import from `@dawn-ai/sdk/testing` instead. This subpath
+ * re-exports the same helpers for back-compat and will continue to work,
+ * but new code should prefer the canonical SDK home.
+ */
+export { expectError, expectMeta, expectOutput } from "@dawn-ai/sdk/testing"
